Bootstrap Navbar
Hello everybody, today I just want to share with all of you  How i created a website with a navbar
using HTML CSS Jquery Bootstrap
what's  bootstrap Navbar? The navbar is the structural element that contains the links that allow you to move between the various sections of the website: in a nutshell it is the area of the page that houses the menu and other utilities
in this site I used stylesheet from https://cdnjs.cloudflare.com and on the link as you can see the first I have linked to eBay.it and another link it's not connected so you can modify it later, also with the logo, as you can see I did not change the logo but u can modify it too, the code is in navbar.html and style.css  hopefully it's helpful to create a website faster.